refactor(tests/system/{common,nar-info-querying}): extract hash and content gen utils
